+++ packages/babel-preset-expo/src/use-dom-directive-plugin.ts
@@ -6,7 +6,7 @@ import crypto from 'crypto';
 import { basename } from 'path';
 import url from 'url';
 
-import { getIsProd } from './common';
+import { getIsProd, getPossibleProjectRoot } from './common';
 
 export function expoUseDomDirectivePlugin(
   api: ConfigAPI & { types: typeof types }
@@ -16,6 +16,7 @@ export function expoUseDomDirectivePlugin(
   // TODO: Is exporting
   const isProduction = api.caller(getIsProd);
   const platform = api.caller((caller) => (caller as any)?.platform);
+  const projectRoot = api.caller(getPossibleProjectRoot);
 
   return {
     name: 'expo-use-dom-directive',
@@ -76,6 +77,29 @@ export function expoUseDomDirectivePlugin(
           );
         }
 
+        // Assert that _layout routes cannot be used in DOM components.
+        const fileBasename = basename(filePath);
+
+        if (
+          projectRoot &&
+          // Detecting if the file is in the router root would be extensive as it would cause a more complex
+          // cache key for each file. Instead, let's just check if the file is in the project root and is not a node_module,
+          // then we can assert that users should not use `_layout` or `+api` with "use dom".
+          filePath.includes(projectRoot) &&
+          !filePath.match(/node_modules/)
+        ) {
+          if (fileBasename.match(/^_layout\.[jt]sx?$/)) {
+            throw path.buildCodeFrameError(
+              'Layout routes cannot be marked as DOM components because they cannot render native views.'
+            );
+          } else if (
+            // No API routes
+            fileBasename.match(/\+api\.[jt]sx?$/)
+          ) {
+            throw path.buildCodeFrameError('API routes cannot be marked as DOM components.');
+          }
+        }
+
         const outputKey = url.pathToFileURL(filePath).href;
 
         const proxyModule: string[] = [
@@ -100,7 +124,7 @@ export function expoUseDomDirectivePlugin(
         } else {
           proxyModule.push(
             // Add the basename to improve the Safari debug preview option.
-            `const source = { uri: new URL("/_expo/@dom/${basename(filePath)}?file=" + ${JSON.stringify(outputKey)}, require("react-native/Libraries/Core/Devtools/getDevServer")().url).toString() };`
+            `const source = { uri: new URL("/_expo/@dom/${fileBasename}?file=" + ${JSON.stringify(outputKey)}, require("react-native/Libraries/Core/Devtools/getDevServer")().url).toString() };`
           );
         }